o Code cleanup

+++ MapView/Map/RMMemoryCache.h
@@ -32,11 +32,12 @@
 @interface RMMemoryCache : NSObject<RMTileCache> {
 	NSMutableDictionary *cache;
 
-	int capacity;
+	NSUInteger capacity;
 }
 
--(id)initWithCapacity: (NSUInteger) _capacity;
+- (id)initWithCapacity:(NSUInteger)_capacity;
+
 /// Remove the least-recently used image from cache, if cache is at or over capacity. Removes only 1 image.
--(void)makeSpaceInCache;
+- (void)makeSpaceInCache;
 
 @end
